Cors오류 수정/상세 기능들 오류 수정 #1

+++ src/main/java/com/example/team3_miniproject/service/MemeService.java
@@ -4,7 +4,7 @@ import com.example.team3_miniproject.dto.AnswerRequestDto;
 import com.example.team3_miniproject.dto.MemeListResponseDto;
 import com.example.team3_miniproject.dto.MemeRequestDto;
 import com.example.team3_miniproject.dto.MemeResponseDto;
-import com.example.team3_miniproject.entity.AnswerReply;
+import com.example.team3_miniproject.entity.Answer;
 import com.example.team3_miniproject.entity.MemeBoard;
 import com.example.team3_miniproject.entity.User;
 import com.example.team3_miniproject.repository.AnswerReplyRepository;
@@ -15,15 +15,11 @@ import com.example.team3_miniproject.repository.MemeRepository;
 import com.example.team3_miniproject.repository.UserRepository;
 import com.example.team3_miniproject.s3.S3Uploader;
 import lombok.RequiredArgsConstructor;
-import org.h2.api.ErrorCode;
-import org.springframework.data.domain.Page;
-import org.springframework.data.domain.Pageable;
 import org.springframework.data.domain.Sort;
 import org.springframework.stereotype.Service;
 import org.springframework.transaction.annotation.Transactional;
 import org.springframework.web.multipart.MultipartFile;
 import java.io.IOException;
-import java.util.ArrayList;
 import java.util.List;
 import java.util.stream.Collectors;
 
@@ -34,7 +30,6 @@ public class MemeService {
 
     private final MemeRepository memeRepository;
     private final AnswerRepository answerRepository;
-
     private final AnswerReplyRepository answerReplyRepository;
     private final UserRepository userRepository;
     private final S3Uploader s3Uploader;
@@ -43,11 +38,13 @@ public class MemeService {
     // 밈 게시물 작성
     // 작성자 : 박소연, 수정자 : 김규리(12/20)
     public MemeResponseDto saveMeme(MemeRequestDto requestDto, User user, MultipartFile multipartFile, String dirName ) throws IOException {
+        System.out.println("밈 게시물 작성 서비스 로직 시작");
         MemeBoard meme;
 
         checkUserExists(userRepository, user);
         String attachedFiles = s3Uploader.upload(multipartFile, dirName);                 // 사진 업로드 (s3 Uploader)
         meme = memeRepository.save(requestDto.toEntity(attachedFiles, user));             // 업로드 파일 경로 + RequestDto(+User) DB에 저장
+        System.out.println("밈 게시물 작성 서비스 로직 끝");
         return new MemeResponseDto(meme);
     }
 
@@ -63,18 +60,12 @@ public class MemeService {
         MemeBoard memeBoard = memeRepository.findById(id).orElseThrow(
                 () -> new RequestException(ErrorCode.NOT_FOUND_BOARD_404)                  // 확인할 게시글이 없습니다.
         );
-        List<AnswerReply> replies = answerReplyRepository.findAllByMemeBoard(memeBoard);
-
-        if (replies.isEmpty()){
-            return new MemeResponseDto(memeBoard);
-        } else {
-            return new MemeResponseDto(memeBoard);
-        }
+        return new MemeResponseDto(memeBoard);
     }
 
     // 밈 게시글 수정
     @Transactional
-    public MemeResponseDto updateMeme(Long id, MemeRequestDto memeRequestDto, MultipartFile multipartFile, String dirName ) throws IOException {
+    public MemeResponseDto updateMeme(Long id, MemeRequestDto memeRequestDto, MultipartFile multipartFile, String dirName, User user) throws IOException {
         // id와 일치하는 게시글 유무
         MemeBoard memeBoard = memeRepository.findById(id).orElseThrow(
                 () -> new RequestException(ErrorCode.NOT_FOUND_BOARD_404)                   // 수정할 밈 게시글이 없습니다.
@@ -83,7 +74,7 @@ public class MemeService {
         String attachedFiles = s3Uploader.upload(multipartFile, dirName);
 
         // 게시글 작성자 확인
-        if (memeBoard.getUsername().equals(memeRequestDto.getUsername())) {
+        if (memeBoard.getUsername().equals(user.getUsername())) {
             // 게시글 업데이트
             memeBoard.update(memeRequestDto,attachedFiles);
             // 게시글 정답정보 삭제
@@ -95,13 +86,16 @@ public class MemeService {
     }
 
     @Transactional
-    public MemeResponseDto incollectAnswer(Long id, AnswerRequestDto request){
+    public MemeResponseDto incollectAnswer(Long id, AnswerRequestDto request, User user){
 
         MemeBoard memeBoard = memeRepository.findById(id).orElseThrow(
                 () -> new RequestException(ErrorCode.NOT_FOUND_BOARD_404)                   // 게시물이 존재하지 않습니다.
         );
 
+        Answer answer;
+
         if (memeBoard.getAnswerValue() == request.getAnswerValue()){
+//            answer = answerRepository.save();
             memeBoard.statusUpdate(true);
             return new MemeResponseDto(memeBoard);
         } else {
@@ -113,8 +107,13 @@ public class MemeService {
     public void deleteMeme(Long id, User user) {
         checkUserExists(userRepository, user);
         MemeBoard meme = checkMemeBoardExists(memeRepository, id);
-        memeRepository.delete(meme);
-        answerRepository.deleteByMemeBoard(meme);
+
+        if (meme.getUsername().equals(user.getUsername())) {
+            memeRepository.delete(meme);
+            answerRepository.deleteByMemeBoard(meme);
+        } else {
+            throw new RequestException(ErrorCode.NULL_USER_ACCESS_403);                      // 게시글 작성자가 아닙니다.
+        }
     }
 
     private User checkUserExists(UserRepository userRepository, User user) {
